requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
toml==0.10.2
loguru==0.7.2
httpx==0.26.0
pydantic==2.6.1 